cleanup(plugin): move list_tables impl to tables::import
