+++ falco_plugin/src/tables/import/table_input.rs
@@ -2,7 +2,7 @@ use crate::error::as_result::WithLastError;
 use crate::tables::import::traits::{TableAccess, TableMetadata};
 use crate::tables::import::RawTable;
 use crate::tables::{Key, TablesInput};
-use falco_plugin_api::ss_plugin_state_type;
+use falco_plugin_api::{ss_plugin_state_type, ss_plugin_table_info};
 use std::ffi::CStr;
 
 impl TablesInput<'_> {
@@ -29,6 +29,21 @@ impl TablesInput<'_> {
             let table = RawTable { table };
             let metadata = T::Metadata::new(&table, self)?;
             Ok(T::new(table, metadata, false))
+        }
+    }
+
+    /// # List the available tables
+    ///
+    /// **Note**: this method is of limited utility in actual plugin code (you know the tables you
+    /// want to access), so it returns the unmodified structure from the plugin API, including
+    /// raw pointers to C-style strings. This may change later.
+    pub fn list_tables(&self) -> &[ss_plugin_table_info] {
+        let mut num_tables = 0u32;
+        let tables = unsafe { (self.list_tables)(self.owner, &mut num_tables as *mut _) };
+        if tables.is_null() {
+            &[]
+        } else {
+            unsafe { std::slice::from_raw_parts(tables, num_tables as usize) }
         }
     }
 }
